+++ src/NeoComposeUnity/Tests/P71ListRepeatTests.cs
@@ -359,7 +359,7 @@ namespace NeoCompose.Tests
                 function = new CountFunction
                 {
                     type = FunctionKind.Count,
-                    info = new FunctionCollectionInfo
+                    info = new FunctionCollectionOptionalBoolInfo
                     {
                         collectionPointer = ListLiteral(values),
                     },